Read settings file when launching TUI (#2)
