+++ src/settings/mod.rs
@@ -12,193 +12,59 @@
 // You should have received a copy of the GNU General Public License along with
 // Cicero. If not, see <https://www.gnu.org/licenses/>.
 
-use std::collections::BTreeSet;
+#[allow(clippy::module_inception)] // Can't think of a better name than "settings"
+mod settings;
+pub use settings::Settings;
+
+pub fn get_settings() -> Settings {
+    match read_settings_file() {
+        Ok(settings) => settings,      // Return the settings read from file
+        Err(_) => Settings::default(), // Failed to read the settings file, return default settings
+    }
+}
+
+use std::error;
 use std::fmt;
-use std::path::PathBuf;
+use std::fs::File;
+use std::io::Read;
 
-use serde::de::Visitor;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use unic::ucd::BlockIter;
+const SETTINGS_FILE_PATH_COMPONENT_CONFIG: &str = ".config";
+const SETTINGS_FILE_PATH_COMPONENT_CICERO: &str = "cicero";
+const SETTINGS_FILE_PATH_COMPONENT_SETTINGS: &str = "settings.toml";
 
-use crate::ucd::{code_point_to_string, string_to_code_point, Plane};
-
-#[derive(Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct Settings {
-    use_fontconfig: bool,
-    font_search_paths: Vec<PathBuf>,
-    preview_fonts: Vec<PreviewFontSetting>,
+#[derive(Debug)]
+enum Error {
+    FailedToGetHomeDir,
 }
 
-#[derive(Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct PreviewFontSetting {
-    code_point_range: Option<CodePointRange>,
-    font_name: String,
-}
-
-const CODE_POINT_RANGE_SEPARATOR: &str = "..";
-
-#[derive(Debug, Eq, PartialEq)]
-enum CodePointRange {
-    Raw { first: char, last: char },
-    Plane { name: String },
-    Block { name: String },
-}
-
-impl Serialize for CodePointRange {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
+impl fmt::Display for Error {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            CodePointRange::Raw { first, last } => serializer.serialize_str(&format!(
-                "{}{}{}",
-                code_point_to_string(*first),
-                CODE_POINT_RANGE_SEPARATOR,
-                code_point_to_string(*last)
-            )),
-            CodePointRange::Plane { name } => serializer.serialize_str(name),
-            CodePointRange::Block { name } => serializer.serialize_str(name),
+            Error::FailedToGetHomeDir => write!(f, "Unable to get home directory"),
         }
     }
 }
 
-impl<'de> Deserialize<'de> for CodePointRange {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct CodePointRangeVisitor {
-            plane_names: BTreeSet<&'static str>,
-            block_names: BTreeSet<&'static str>,
-        }
+impl error::Error for Error {}
 
-        impl CodePointRangeVisitor {
-            fn new() -> Self {
-                let plane_names = {
-                    let mut set = BTreeSet::new();
-                    for name in Plane::all_plane_names() {
-                        set.insert(*name);
-                    }
-                    set
-                };
+type Result<T> = std::result::Result<T, Box<dyn error::Error>>;
 
-                let block_names = {
-                    let block_iter = BlockIter::new();
-                    let mut set = BTreeSet::new();
-                    for name in block_iter {
-                        set.insert(name.name);
-                    }
-                    set
-                };
+fn read_settings_file() -> Result<Settings> {
+    let settings_path = {
+        let mut home_path = dirs::home_dir().ok_or(Error::FailedToGetHomeDir)?;
+        home_path.push(SETTINGS_FILE_PATH_COMPONENT_CONFIG);
+        home_path.push(SETTINGS_FILE_PATH_COMPONENT_CICERO);
+        home_path.push(SETTINGS_FILE_PATH_COMPONENT_SETTINGS);
+        home_path
+    };
 
-                CodePointRangeVisitor {
-                    plane_names,
-                    block_names,
-                }
-            }
-        }
+    let mut file = File::open(settings_path)?;
 
-        impl<'de> Visitor<'de> for CodePointRangeVisitor {
-            type Value = CodePointRange;
+    let mut file_content = String::new();
+    file.read_to_string(&mut file_content)?;
 
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str(
-                    "A string representing either an inclusive range, a plane name, or a block name")
-            }
-
-            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                let mut range_components = v.split(CODE_POINT_RANGE_SEPARATOR);
-                let first_str = range_components.next();
-                let last_str = range_components.next();
-                if first_str.is_some() && last_str.is_some() {
-                    let first =
-                        string_to_code_point(first_str.unwrap()).ok_or(E::custom(format!(
-                            "Invalid first code point for range: '{}'",
-                            first_str.unwrap()
-                        )))?;
-                    let last = string_to_code_point(last_str.unwrap()).ok_or(E::custom(
-                        format!("Invalid last code point for range: '{}'", last_str.unwrap()),
-                    ))?;
-                    return Ok(CodePointRange::Raw { first, last });
-                }
-
-                if self.plane_names.contains(v) {
-                    return Ok(CodePointRange::Plane { name: v.to_owned() });
-                }
-
-                if self.block_names.contains(v) {
-                    return Ok(CodePointRange::Block { name: v.to_owned() });
-                }
-
-                Err(E::custom(format!("Unrecognized code point range: '{}'", v)))
-            }
-        }
-
-        deserializer.deserialize_str(CodePointRangeVisitor::new())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    const TEST_SETTINGS_TOML_STRING: &str = include_str!("test_resources/test_settings.toml");
-
-    fn get_test_settings() -> Settings {
-        Settings {
-            use_fontconfig: true,
-            font_search_paths: vec![
-                PathBuf::from("/test/path/fonts"),
-                PathBuf::from("/test/path/fonts2"),
-            ],
-            preview_fonts: vec![
-                PreviewFontSetting {
-                    code_point_range: None,
-                    font_name: "TestFontName-Regular".to_owned(),
-                },
-                PreviewFontSetting {
-                    code_point_range: Some(CodePointRange::Raw {
-                        first: '\u{0020}',
-                        last: '\u{00FF}',
-                    }),
-                    font_name: "TestFontName-Regular".to_owned(),
-                },
-                PreviewFontSetting {
-                    code_point_range: Some(CodePointRange::Block {
-                        name: "Basic Latin".to_owned(),
-                    }),
-                    font_name: "TestFontName-Regular".to_owned(),
-                },
-                PreviewFontSetting {
-                    code_point_range: Some(CodePointRange::Plane {
-                        name: "Basic Multilingual Plane".to_owned(),
-                    }),
-                    font_name: "TestFontName-Regular".to_owned(),
-                },
-            ],
-        }
-    }
-
-    fn sanitize_line_breaks(input: &str) -> String {
-        input.replace("\r\n", "\n")
-    }
-
-    #[test]
-    fn test_serialization() {
-        let serialized_toml_string = toml::to_string(&get_test_settings()).unwrap();
-        assert_eq!(
-            sanitize_line_breaks(&serialized_toml_string),
-            sanitize_line_breaks(TEST_SETTINGS_TOML_STRING)
-        );
-    }
-
-    #[test]
-    fn test_deserialization() {
-        let deserialized_settings: Settings =
-            toml::from_str(&sanitize_line_breaks(TEST_SETTINGS_TOML_STRING)).unwrap();
-        assert_eq!(deserialized_settings, get_test_settings());
+    match toml::from_str(&file_content) {
+        Ok(settings) => Ok(settings),
+        Err(error) => Err(Box::new(error)),
     }
 }
